requirements.txt
beautifulsoup4==4.12.2
boto3==1.33.6
numpy==1.26.2
openai==1.3.8
pandas>=2.0.0
Pillow==9.5.0
Requests==2.31.0
sqlalchemy==1.4.49
streamlit==1.26.0
snowflake-sqlalchemy
python-dotenv==1.0.0
torch
clip
fastapi
uvicorn